+++ mulighetsrommet-api/src/main/kotlin/no/nav/mulighetsrommet/api/clients/arenaadapter/ArenaAdapterClientImpl.kt
@@ -7,7 +7,7 @@ import io.ktor.client.plugins.*
 import io.ktor.client.plugins.cache.*
 import io.ktor.client.request.*
 import io.ktor.http.*
-import no.nav.mulighetsrommet.api.setup.http.httpJsonClient
+import no.nav.mulighetsrommet.ktor.clients.httpJsonClient
 import no.nav.mulighetsrommet.domain.dto.ArenaTiltaksgjennomforingsstatusDto
 import no.nav.mulighetsrommet.domain.dto.ExchangeArenaIdForIdResponse
 import org.slf4j.LoggerFactory
